+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/notifications/NotificationServiceTest.scala
@@ -8,7 +8,7 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatestplus.mockito.MockitoSugar
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.deployment.{DeployedScenarioData, DeploymentManager}
-import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName}
+import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.{DeploymentData, DeploymentId, ExternalDeploymentId}
@@ -21,11 +21,12 @@ import pl.touk.nussknacker.ui.api.helpers.TestCategories.TestCat
 import pl.touk.nussknacker.ui.api.helpers.TestProcessingTypes.Streaming
 import pl.touk.nussknacker.ui.api.helpers.{TestFactory, WithHsqlDbTesting}
 import pl.touk.nussknacker.ui.db.entity.ProcessActionId
+import pl.touk.nussknacker.ui.listener.ProcessChangeListener
 import pl.touk.nussknacker.ui.process.deployment.LoggedUserConversions._
 import pl.touk.nussknacker.ui.process.deployment.{DeploymentManagerDispatcher, DeploymentServiceImpl, ScenarioResolver}
 import pl.touk.nussknacker.ui.process.processingtypedata.MapBasedProcessingTypeDataProvider
 import pl.touk.nussknacker.ui.process.repository.ProcessRepository.CreateProcessAction
-import pl.touk.nussknacker.ui.process.repository.{DbProcessActionRepository, RepositoryManager}
+import pl.touk.nussknacker.ui.process.repository.{DBIOActionRunner, DbProcessActionRepository}
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.util.DBIOActionValues
 import pl.touk.nussknacker.ui.validation.ProcessValidation
@@ -38,7 +39,7 @@ import scala.util.{Failure, Success, Try}
 
 class NotificationServiceTest extends AnyFunSuite with Matchers with PatientScalaFutures with MockitoSugar with WithHsqlDbTesting with EitherValuesDetailedMessage with OptionValues with DBIOActionValues {
 
-  protected val repositoryManager: RepositoryManager = RepositoryManager.createDbRepositoryManager(db)
+  override protected val dbioRunner: DBIOActionRunner = DBIOActionRunner(db)
 
   private var currentInstant: Instant = Instant.ofEpochMilli(0)
   private val clock: Clock = clockForInstant(() => currentInstant)
@@ -54,18 +55,16 @@ class NotificationServiceTest extends AnyFunSuite with Matchers with PatientScal
     val id = saveSampleProcess(processName)
     val processIdWithName = ProcessIdWithName(id, processName)
 
-    val listener = new NotificationsListener(NotificationConfig(20 minutes), (_: ProcessId) => Future.successful(Some(processName)), clock)
     val deploymentManager = mock[DeploymentManager]
-    val deploymentService = createDeploymentService(listener, deploymentManager)
-    val notificationService = new ListenerBasedNotificationService(listener)
+    val (deploymentService, notificationService) = createServices(deploymentManager, processName)
     def notificationsFor(user: LoggedUser, after: Option[Instant] = None): List[Notification] = notificationService.notifications(after)(user, ctx).futureValue
     def deployProcess(givenDeployResult: Try[Option[ExternalDeploymentId]], user: LoggedUser): Option[ExternalDeploymentId] = {
       when(deploymentManager.deploy(any[ProcessVersion], any[DeploymentData], any[CanonicalProcess], any[Option[String]])).thenReturn(Future.fromTry(givenDeployResult))
       deploymentService.deployProcessAsync(processIdWithName, None, None)(user, ctx).flatten.futureValue
     }
 
-    val userForSuccess = TestFactory.adminUser(id = "successUser")
-    val userForFail = TestFactory.adminUser("failUser")
+    val userForSuccess = TestFactory.adminUser("successUser", "successUser")
+    val userForFail = TestFactory.adminUser("failUser", "failUser")
 
     deployProcess(Success(None), userForSuccess)
     notificationsFor(userForSuccess).map(_.toRefresh) shouldBe List(expectedRefreshAfterSuccess)
@@ -83,15 +82,18 @@ class NotificationServiceTest extends AnyFunSuite with Matchers with PatientScal
     notificationsFor(userForFail).map(_.toRefresh) shouldBe Symbol("empty")
   }
 
-  private def createDeploymentService(listener: NotificationsListener, deploymentManager: DeploymentManager) = {
+  private def createServices(deploymentManager: DeploymentManager, processName: ProcessName) = {
     when(deploymentManager.findJobStatus(any[ProcessName])).thenReturn(Future.successful(None))
     val managerDispatcher = mock[DeploymentManagerDispatcher]
     when(managerDispatcher.deploymentManager(any[String])).thenReturn(deploymentManager)
-    new DeploymentServiceImpl(managerDispatcher, processRepository, actionRepository, repositoryManager, mock[ProcessValidation], mock[ScenarioResolver], listener, clock) {
+    val config = NotificationConfig(20 minutes)
+    val notificationService = new NotificationServiceImpl(actionRepository, dbioRunner, config, clock)
+    val deploymentService = new DeploymentServiceImpl(managerDispatcher, processRepository, actionRepository, dbioRunner, mock[ProcessValidation], mock[ScenarioResolver], mock[ProcessChangeListener], clock) {
       override protected def validateBeforeDeploy(processDetails: processdetails.BaseProcessDetails[CanonicalProcess], actionId: ProcessActionId)(implicit user: LoggedUser, ec: ExecutionContext): Future[DeployedScenarioData] = {
         Future.successful(DeployedScenarioData(processDetails.toEngineProcessVersion, prepareDeploymentData(user.toManagerUser, DeploymentId(actionId.toString)), processDetails.json))
       }
     }
+    (deploymentService, notificationService)
   }
 
   private def saveSampleProcess(processName: ProcessName) = {